// src/api-client/model/displayCalendarScheduleDetail.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * 빌릴게(Billilge) API
 * 국민대학교 소프트웨어융합대학 복지물품 대여 서비스
 * OpenAPI spec version: v1
 */

export interface DisplayCalendarScheduleDetail {
  /** 일정 ID */
  scheduleId: number;
  /** 일정 날짜 */
  date: string;
  /** 일정 목록 */
  schedules: string[];
}