// src/api-client/model/displayPosterDetail.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * 빌릴게(Billilge) API
 * 국민대학교 소프트웨어융합대학 복지물품 대여 서비스
 * OpenAPI spec version: v1
 */

export interface DisplayPosterDetail {
  /** 포스터 ID */
  posterId: number;
  /** 포스터 제목 */
  title: string;
  /** 포스터 이미지 URL */
  imageUrl: string;
  /** 활성화 여부 */
  isActive: boolean;
  /** 생성일시 */
  createdAt: string;
}